Make script directly executable so can run directly from source
1. Change FROM original concept of a single file TO a directory
FROM using `make` to build a single shell
  script with all package manager plugin scripts concatenated into a
  single script file
TO a directly with each plugin file remaining separate.

This new approach will be more performant as only the plugin for the
current system will be loaded.  Also, plugins can be implemented in
languages other than POSIX shell.

2. Find the actual installed directory rather than where the original
calling point.
Ex: If 'usm' symlinked to to ~/bin, then can now find the plugins in the
actual installed directory.